requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
apscheduler==3.10.4
cachetools==5.3.2